+++ packages/react-ui/src/components/AccountVerify/AccountVerify.tsx
@@ -1,7 +1,7 @@
 import React, {useEffect, useState} from 'react';
 import {styled} from '@mui/material/styles';
 import {SCUserContextType, useSCUser} from '@selfcommunity/react-core';
-import {CircularProgress, Typography} from '@mui/material';
+import {Alert, Box, CircularProgress} from '@mui/material';
 import classNames from 'classnames';
 import {useThemeProps} from '@mui/system';
 import {AccountService} from '@selfcommunity/api-services';
@@ -15,7 +15,7 @@ const classes = {
   error: `${PREFIX}-error`
 };
 
-const Root = styled(Typography, {
+const Root = styled(Box, {
   name: PREFIX,
   slot: 'Root',
   overridesResolver: (props, styles) => styles.root
@@ -45,7 +45,7 @@ export interface AccountVerifyProps {
    * Callback triggered on success sign in
    * @default null
    */
-  onSuccess?: () => void;
+  onSuccess?: (res: any) => void;
 
   /**
    * Action component to display after success message
@@ -77,8 +77,8 @@ export interface AccountVerifyProps {
  |Rule Name|Global class|Description|
  |---|---|---|
  |root|.SCAccountVerify-root|Styles applied to the root element.|
- |success|.SCAccountRecover-success|Styles applied to the success Typography.|
- |error|.SCAccountRecover-error|Styles applied to the error Typography.|
+ |success|.SCAccountVerify-success|Styles applied to the success Alert.|
+ |error|.SCAccountVerify-error|Styles applied to the error Alert.|
 
  *
  * @param inProps
@@ -94,6 +94,7 @@ export default function AccountVerify(inProps: AccountVerifyProps): JSX.Element 
 
   // STATE
   const [isValidating, setIsValidating] = useState<boolean>(true);
+  const [succeed, setSucceed] = useState<boolean | string>(false);
   const [error, setError] = useState<boolean>(false);
 
   // CONTEXT
@@ -104,7 +105,10 @@ export default function AccountVerify(inProps: AccountVerifyProps): JSX.Element 
     setIsValidating(true);
     setError(false);
     AccountService.verify({validation_code: validationCode})
-      .then(() => onSuccess && onSuccess())
+      .then((res: any) => {
+        setSucceed(res.token ? true : res.user_awaiting_approval);
+        onSuccess && onSuccess(res);
+      })
       .catch(() => setError(true))
       .then(() => setIsValidating(false));
   }, [validationCode]);
@@ -116,18 +120,20 @@ export default function AccountVerify(inProps: AccountVerifyProps): JSX.Element 
 
   // RENDER
   return (
-    <Root className={classNames(classes.root, className, {[classes.error]: error, [classes.success]: !isValidating && !error})} {...rest}>
-      {isValidating ? (
+    <Root className={classNames(classes.root, className)} {...rest}>
+      {succeed ? (
+        <Alert severity="success" className={classes.success}>
+          {typeof succeed === 'string' ? succeed : <FormattedMessage id="ui.accountVerify.success" defaultMessage="ui.accountVerify.success" />}
+          {successAction}
+        </Alert>
+      ) : error ? (
+        <Alert severity="error" className={classes.error}>
+          <FormattedMessage id="ui.accountVerify.error" defaultMessage="ui.accountVerify.error" />
+        </Alert>
+      ) : (
         <>
           <CircularProgress />
           <FormattedMessage id="ui.accountVerify.verifying" defaultMessage="ui.accountVerify.verifying" />
-        </>
-      ) : error ? (
-        <FormattedMessage id="ui.accountVerify.error" defaultMessage="ui.accountVerify.error" />
-      ) : (
-        <>
-          <FormattedMessage id="ui.accountVerify.success" defaultMessage="ui.accountVerify.success" />
-          {successAction}
         </>
       )}
     </Root>