+++ packages/react-ui/src/components/Header/MobileHeader/MobileHeader.tsx
@@ -1,6 +1,6 @@
 import {AppBar, Badge, Box, Button, Grid, IconButton, styled, SwipeableDrawer, Tab, Tabs, Toolbar} from '@mui/material';
 import Icon from '@mui/material/Icon';
-import React, {useContext, useState} from 'react';
+import React, {useContext, useEffect, useState} from 'react';
 import {Link, SCPreferences, SCUserContext, SCUserContextType, useSCPreferences} from '@selfcommunity/react-core';
 import {useThemeProps} from '@mui/system';
 import SearchBar, {HeaderSearchBarProps} from '../SearchBar';
@@ -8,13 +8,13 @@ import classNames from 'classnames';
 import {FormattedMessage} from 'react-intl';
 import HeaderMenu from '../HeaderMenu';
 import {SCHeaderMenuUrlsType} from '../../../types';
+import MobileHeaderSkeleton from './Skeleton';
 
 const PREFIX = 'SCMobileHeader';
 
 const classes = {
   root: `${PREFIX}-root`,
   tabs: `${PREFIX}-tabs`,
-  settingsTab: `${PREFIX}-setting-tab`,
   topToolbar: `${PREFIX}-top-toolbar`,
   bottomToolbar: `${PREFIX}-bottom-toolbar`
 };
@@ -27,13 +27,6 @@ const Root = styled(Box, {
   [`& .${classes.tabs}`]: {
     '& .MuiTabs-indicator': {
       top: '0px'
-    }
-  },
-  [`& .${classes.settingsTab}`]: {
-    '& .MuiTabs-indicator': {
-      top: '0px',
-      backgroundColor: '#000000',
-      opacity: 0
     }
   },
   [`& .${classes.topToolbar}`]: {
@@ -107,11 +100,26 @@ export default function MobileHeader(inProps: MobileHeaderProps) {
     }
   };
 
+  useEffect(() => {
+    const getSelectedTab = JSON.parse(localStorage.getItem('selectedTab'));
+    if (getSelectedTab) {
+      setValue(getSelectedTab);
+    }
+  }, []);
+
+  useEffect(() => {
+    localStorage.setItem('selectedTab', JSON.stringify(value));
+  }, [value]);
+
+  if (scUserContext.loading) {
+    return <MobileHeaderSkeleton />;
+  }
+
   return (
-    <Root className={classNames(classes.root, className)}>
-      <AppBar position="fixed" color={'default'}>
+    <Root className={classNames(classes.root, className)} {...rest}>
+      <AppBar position='fixed' color={'default'}>
         <Toolbar className={classes.topToolbar}>
-          <Grid container direction="row" justifyContent="flex-start">
+          <Grid container direction='row' justifyContent='flex-start'>
             {scUserContext.user && url && url.home ? (
               <Link to={url.home}>
                 <img src={logo} alt={'logo'} style={{height: '30px'}} />
@@ -124,7 +132,7 @@ export default function MobileHeader(inProps: MobileHeaderProps) {
           </Grid>
           <SearchBar {...searchBarProps} />
           {scUserContext.user && url && url.create && (
-            <IconButton component={Link} to={url.create} size="large" aria-label="New Contribute" color="inherit">
+            <IconButton component={Link} to={url.create} size='large' aria-label='New Contribute' color='inherit'>
               <Icon>create</Icon>
             </IconButton>
           )}
@@ -134,53 +142,54 @@ export default function MobileHeader(inProps: MobileHeaderProps) {
         {scUserContext.user ? (
           <Toolbar className={classes.bottomToolbar}>
             <Tabs
-              className={value === 4 ? classes.settingsTab : classes.tabs}
+              className={classes.tabs}
               onChange={(e, v) => setValue(v)}
               value={value}
-              textColor="primary"
-              indicatorColor="primary"
-              aria-label="Navigation Tabs"
-              variant="scrollable">
-              {url && url.home && <Tab icon={<Icon>home</Icon>} aria-label="HomePage" to={url.home} component={Link}></Tab>}
-              {url && url.explore && <Tab icon={<Icon>explore</Icon>} aria-label="Explore" to={url.explore} component={Link}></Tab>}
-              {url && url.followings && <Tab icon={<Icon>person</Icon>} aria-label="Followings" to={url.followings} component={Link}></Tab>}
+              textColor='primary'
+              indicatorColor='primary'
+              aria-label='Navigation Tabs'
+              variant='scrollable'>
+              {url && url.home && <Tab value={0} icon={<Icon>home</Icon>} aria-label="HomePage" to={url.home} component={Link}></Tab>}
+              {url && url.explore && <Tab value={1} icon={<Icon>explore</Icon>} aria-label="Explore" to={url.explore} component={Link}></Tab>}
+              {url && url.followings && <Tab value={2} icon={<Icon>person</Icon>} aria-label="Followings" to={url.followings} component={Link}></Tab>}
               {url && url.notifications && (
                 <Tab
+                  value={3}
                   icon={
-                    <Badge badgeContent={scUserContext.user.unseen_interactions_counter} color="error">
+                    <Badge badgeContent={scUserContext.user.unseen_interactions_counter} color='error'>
                       <Icon>notifications</Icon>
                     </Badge>
                   }
-                  aria-label="Notifications"
+                  aria-label='Notifications'
                   to={url.notifications}
                   component={Link}></Tab>
               )}
-              <Tab className={classes.settingsTab} icon={<Icon>menu</Icon>} aria-label="HeaderMenu" onClick={handleOpenSettingsMenu}></Tab>
-              <SwipeableDrawer
-                anchor={'right'}
-                open={openSettings}
-                onClick={() => setOpenSettings(false)}
-                onClose={() => setOpenSettings(false)}
-                onOpen={toggleDrawer('right', true)}>
-                <HeaderMenu url={url} />
-              </SwipeableDrawer>
             </Tabs>
+            <IconButton onClick={handleOpenSettingsMenu}><Icon>menu</Icon></IconButton>
+            <SwipeableDrawer
+              anchor={'right'}
+              open={openSettings}
+              onClick={() => setOpenSettings(false)}
+              onClose={() => setOpenSettings(false)}
+              onOpen={toggleDrawer('right', true)}>
+              <HeaderMenu onItemClick={()=> setValue(null)} url={url} />
+            </SwipeableDrawer>
           </Toolbar>
         ) : (
           <Toolbar sx={{justifyContent: 'space-between'}}>
             {url && url.explore && (
-              <Button component={Link} to={url.explore} size="medium" color="inherit">
-                <FormattedMessage id="ui.header.button.explore" defaultMessage="ui.header.button.explore" />
+              <Button component={Link} to={url.explore} size='medium' color='inherit'>
+                <FormattedMessage id='ui.header.button.explore' defaultMessage='ui.header.button.explore' />
               </Button>
             )}
             {url && url.login && (
-              <Button color="inherit" onClick={url.login}>
-                <FormattedMessage id="ui.header.button.login" defaultMessage="ui.header.button.login" />
+              <Button color='inherit' onClick={url.login}>
+                <FormattedMessage id='ui.header.button.login' defaultMessage='ui.header.button.login' />
               </Button>
             )}
             {url && url.register && (
-              <Button color="inherit" component={Link} to={url.register}>
-                <FormattedMessage id="ui.header.button.register" defaultMessage="ui.header.button.register" />
+              <Button color='inherit' component={Link} to={url.register}>
+                <FormattedMessage id='ui.header.button.register' defaultMessage='ui.header.button.register' />
               </Button>
             )}
           </Toolbar>