+++ packages/react-ui/src/components/Header/MobileHeader/index.tsx
@@ -1,3 +1,4 @@
 import MobileHeader, {MobileHeaderProps} from './MobileHeader';
+import MobileHeaderSkeleton from './Skeleton';
 export default MobileHeader;
-export {MobileHeaderProps};
+export {MobileHeaderProps, MobileHeaderSkeleton};